requirements.txt
entsoe-py
tensorflow
multiprocessing
statsmodels